requirements.txt
torch
torchvision
timm
Pillow
PyQt5
